requirements.txt
psutil==5.9.8
requests==2.31.0
PyYAML==6.0.1
python-dotenv==1.0.1
